+++ src/process_inspector/__init__.py
@@ -1,5 +1,6 @@
 from .appcontrol import NativeApp  # noqa: F401
 from .oscontrol import OperatingSystem  # noqa: F401
 from .servicecontrol import Service  # noqa: F401
+from .teamviewer import Teamviewer  # noqa: F401
 
 __version__ = "0.1.2"